requirements.txt
numpy>=2.3
scikit-learn==1.7.1
pandas
matplotlib
pyyaml
sentence-transformers
plotly
tqdm
